+++ particular/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased - 2023-20-07
+
+### Added
+
+- `ParticleStorage` trait blanket implemented for storages of `PointMass`.
+- `TreeAffected` storage.
+
+### Changed
+
+- `BarnesHut` compute methods use `TreeAffected` instead of `MassiveAffected`. This gives potential access to the tree built to compute the accelerations.
+
 ## [0.6.1] - 2023-19-07
 
 ### Added